IMP-1 AppContext File added to integrate adminPage into the app context.
